# graph/path: Adjusted bellman ford algorithm to use queue.