--- a/graph/path/bench_test.go
+++ b/graph/path/bench_test.go
@@ -139,3 +139,35 @@ func BenchmarkAStarUndirectedmallWorld_100_5_20_2_Heur(b *testing.B) {
 	}
 	benchmarkAStarHeuristic(b, nswUndirected_100_5_20_2, h)
 }
+
+var (
+	gnpDirected_500_tenth   = gnpDirected(500, 0.1)
+	gnpDirected_1000_tenth  = gnpDirected(1000, 0.1)
+	gnpDirected_2000_tenth 	= gnpDirected(2000, 0.1)
+	gnpDirected_500_half    = gnpDirected(500, 0.5)
+	gnpDirected_1000_half   = gnpDirected(1000, 0.5)
+	gnpDirected_2000_half  	= gnpDirected(2000, 0.5)
+	gnpDirected_500_full    = gnpDirected(500, 1)
+	gnpDirected_1000_full   = gnpDirected(1000, 1)
+	gnpDirected_2000_full  	= gnpDirected(2000, 1)
+)
+
+func gnpDirected(n int, p float64) graph.Directed {
+	g := simple.NewDirectedGraph()
+	gen.Gnp(g, n, p, nil)
+	return g
+}
+
+func benchmarkBellmanFord(b *testing.B, n int, p float64, g graph.Directed) {
+	BellmanFordFrom(g.Node(0), g)
+}
+
+func BenchmarkBellmanFordFrom_500_tenth(b *testing.B) { benchmarkBellmanFord(b, 500, 0.5, gnpDirected_500_tenth) }
+func BenchmarkBellmanFordFrom_1000_tenth(b *testing.B) {benchmarkBellmanFord(b, 1000, 0.5, gnpDirected_1000_tenth) }
+func BenchmarkBellmanFordFrom_2000_tenth(b *testing.B) { benchmarkBellmanFord(b, 2000, 0.5, gnpDirected_2000_tenth) }
+func BenchmarkBellmanFordFrom_500_half(b *testing.B) { benchmarkBellmanFord(b, 500, 0.5, gnpDirected_500_half) }
+func BenchmarkBellmanFordFrom_1000_half(b *testing.B) {benchmarkBellmanFord(b, 1000, 0.5, gnpDirected_1000_half) }
+func BenchmarkBellmanFordFrom_2000_half(b *testing.B) { benchmarkBellmanFord(b, 2000, 0.5, gnpDirected_2000_half) }
+func BenchmarkBellmanFordFrom_500_full(b *testing.B) { benchmarkBellmanFord(b, 500, 1, gnpDirected_500_full) }
+func BenchmarkBellmanFordFrom_1000_full(b *testing.B) {benchmarkBellmanFord(b, 1000, 1, gnpDirected_1000_full) }
+func BenchmarkBellmanFordFrom_2000_full(b *testing.B) { benchmarkBellmanFord(b, 2000, 1, gnpDirected_2000_full) }