Hey there,

I started working on the _todo_ left in the [bellman_ford_moore-file](https://github.com/gonum/gonum/blob/master/graph/path/bellman_ford_moore.go): "Consider adding further optimisations"

This semester I finished an algorithm course at my university. We learned based on [Sedgewick & Wayne](https://books.google.de/books?id=MTpsAQAAQBAJ&pg=PA672&lpg=PA672&dq=Sedgewick+%26+Wayne+queue+bellman+ford&source=bl&ots=QfpHzOG8nZ&sig=ACfU3U32xLZKnJsvoSeKNpQzY_6oTm-GJg&hl=en&sa=X&ved=2ahUKEwjYodKmha3kAhVCKVAKHflyCqwQ6AEwC3oECAkQAQ#v=onepage&q=Sedgewick%20%26%20Wayne%20queue%20bellman%20ford&f=false) suggestion a queue approach is to be preferred since we only have to relax those nodes whose vertex distance changed in the previous iteration.

While the worst case complexity is still **O(|V| . |E|)** it is typically **O(|V| + |E|)**.

I have implemented such queue and I would appreciate any feedback! :)

Best regards,
Dario