@Zarathustra2 2019-09-01
Hey @kortschak,

Thank you very much for reviewing my code and answering so quick!

I have updated all of your requested changes, see latest commit. 

Furthermore, here are the benchmark tests:

Tested with a lenovo e480, 16gb ram. 

Without queue
===
goos: linux
goarch: amd64
pkg: gonum.org/v1/gonum/graph/path

| | | |
| ------------- |:-------------:| -----:|
| BenchmarkBellmanFordFrom_500_tenth-8     |        2000000000        |       0.01 ns/op
BenchmarkBellmanFordFrom_1000_tenth-8     |      2000000000   |            0.02 ns/op
BenchmarkBellmanFordFrom_2000_tenth-8     |      2000000000     |          0.08 ns/op
BenchmarkBellmanFordFrom_500_half-8         |      2000000000      |         0.02 ns/op
BenchmarkBellmanFordFrom_1000_half-8      |       2000000000       |       0.09 ns/op
BenchmarkBellmanFordFrom_2000_half-8      |       2000000000       |        0.38 ns/op
BenchmarkBellmanFordFrom_500_full-8         |       1000000000       |        0.09 ns/op
BenchmarkBellmanFordFrom_1000_full-8       |       2000000000       |        0.19 ns/op
BenchmarkBellmanFordFrom_2000_full-8      |             1                    |        1481679020 ns/op

PASS
ok      gonum.org/v1/gonum/graph/path   102.412s

With queue
=====
goos: linux
goarch: amd64
pkg: gonum.org/v1/gonum/graph/path

| | | |
| ------------- |:-------------:| -----:|
| BenchmarkBellmanFordFrom_500_tenth-8         |   2000000000       |        0.00 ns/op
BenchmarkBellmanFordFrom_1000_tenth-8       |    2000000000      |         0.01 ns/op
BenchmarkBellmanFordFrom_2000_tenth-8       |    1000000000      |         0.06 ns/op
BenchmarkBellmanFordFrom_500_half-8          |   2000000000        |       0.01 ns/op
BenchmarkBellmanFordFrom_1000_half-8        |    2000000000      |         0.03 ns/op
BenchmarkBellmanFordFrom_2000_half-8        |    1000000000      |         0.25 ns/op
BenchmarkBellmanFordFrom_500_full-8           |  2000000000        |       0.02 ns/op
BenchmarkBellmanFordFrom_1000_full-8         |   2000000000       |        0.06 ns/op
BenchmarkBellmanFordFrom_2000_full-8         |   1000000000       |        0.49 ns/op

PASS
ok      gonum.org/v1/gonum/graph/path   78.276s