@kortschak 2019-09-01
Thanks for doing the benchmarking. Would you post the output of `benchstat old.bench new bench` after running the benchmarks with `go test -run ^$ -bench BellmanFordFrom -count 10` redirected appropriately.