@Zarathustra2 2019-09-01
Hey @kortschak,

Thanks again for reviewing! :)

I have fixed all requested changes.

Furthermore, here are the benchmarks. Do you also want me to commit the .txt files with the result? If so, where should I put them?

Benchmark-Results
===============

I had to remove all the tests with 2000 nodes because with the old code a time-out error - _Test killed with quit: ran too long (10m0s)._  - occurred while with the queue approach they all passed.

`benchstat old.txt new.txt`

| | | |
| ------------- |:-------------:| -----:|
| name          |                old time/op | new time/op | delta
| BellmanFordFrom/500_tenth-8  | 13.4ms ± 1%  | 3.5ms ± 1% | -74.18%  (p=0.000 n=10+10)
BellmanFordFrom/1000_tenth-8 | 50.3ms ± 4% | 13.9ms ± 2% | -72.26%  (p=0.000 n=10+10)
BellmanFordFrom/500_half-8  |  46.7ms ± 1% | 15.6ms ± 1% | -66.60%  (p=0.000 n=10+9)
BellmanFordFrom/1000_half-8 |   184ms ± 1% |   61ms ± 1% | -66.60%  (p=0.000 n=9+10)
BellmanFordFrom/500_full-8  |  90.0ms ± 1% | 30.3ms ± 1% | -66.35%  (p=0.000 n=9+10)
BellmanFordFrom/1000_full-8  |  359ms ± 1%  | 121ms ± 1% | -66.31%  (p=0.000 n=10+10)