@kortschak 2019-09-01
Thanks. Here's what I see, which looks good.

Can you send a PR adding yourself to A+C with the commit message "A+C: add Dario Heinisch". Then I'll approve this and merge them. Thank you for a valuable first contrib.

<table class='benchstat'>
<tr><th>name</th><th>old time/op</th><th>new time/op</th><th>delta</th>
<tr><td>BellmanFordFrom/500_tenth-8</td><td>13.0ms ± 2%</td><td>3.5ms ± 7%</td><td>-72.89%</td><td>(p=0.000 n=10+10)</td>
<tr><td>BellmanFordFrom/1000_tenth-8</td><td>53.8ms ± 8%</td><td>13.7ms ± 0%</td><td>-74.56%</td><td>(p=0.000 n=10+9)</td>
<tr><td>BellmanFordFrom/2000_tenth-8</td><td>188ms ±18%</td><td>55ms ± 1%</td><td>-70.82%</td><td>(p=0.000 n=10+10)</td>
<tr><td>BellmanFordFrom/500_half-8</td><td>49.1ms ± 8%</td><td>15.6ms ± 1%</td><td>-68.29%</td><td>(p=0.000 n=10+10)</td>
<tr><td>BellmanFordFrom/1000_half-8</td><td>193ms ± 7%</td><td>62ms ± 0%</td><td>-68.18%</td><td>(p=0.000 n=10+10)</td>
<tr><td>BellmanFordFrom/2000_half-8</td><td>798ms ± 6%</td><td>248ms ± 1%</td><td>-68.95%</td><td>(p=0.000 n=10+10)</td>
<tr><td>BellmanFordFrom/500_full-8</td><td>96.0ms ± 8%</td><td>30.0ms ± 0%</td><td>-68.75%</td><td>(p=0.000 n=10+8)</td>
<tr><td>BellmanFordFrom/1000_full-8</td><td>364ms ± 1%</td><td>121ms ± 1%</td><td>-66.82%</td><td>(p=0.000 n=9+10)</td>
<tr><td>BellmanFordFrom/2000_full-8</td><td>1.50s ± 7%</td><td>0.49s ± 1%</td><td>-67.56%</td><td>(p=0.000 n=10+10)</td>
</table>