@Zarathustra2 2019-09-01
Hey @kortschak,

I have added myself see PR #1070 

Thank you very much, it was really interesting working on this, learned a lot. Also thanks for all the reviews, they helped me a lot!

Hopefully, this will not be last my commit!

Best regards, 
Dario :)